+++ tests/test_listener.py
@@ -1,18 +1,18 @@
 import asyncio
 import contextlib
-from typing import Any, Dict, List, Union
+from typing import Any, Dict, List
 
 import aiopg
 
 import aiopg_listen
 
 
-class Processor:
+class TestHandler:
     def __init__(self, delay: float = 0) -> None:
         self.delay = delay
-        self.notifications: List[Union[aiopg_listen.Notification, aiopg_listen.Timeout]] = []
+        self.notifications: List[aiopg_listen.NotificationOrTimeout] = []
 
-    async def process(self, notification: Union[aiopg_listen.Notification, aiopg_listen.Timeout]) -> None:
+    async def handle(self, notification: aiopg_listen.NotificationOrTimeout) -> None:
         await asyncio.sleep(self.delay)
         self.notifications.append(notification)
 
@@ -24,28 +24,26 @@ async def cancel_and_wait(future: asyncio.Future[None]) -> None:
 
 
 async def test_two_inactive_channels(pg_server: Dict[str, Any]) -> None:
-    processor_1 = Processor()
-    processor_2 = Processor()
-    consumer = aiopg_listen.NotificationConsumer(aiopg_listen.connect_func(**pg_server["pg_params"]))
+    handler_1 = TestHandler()
+    handler_2 = TestHandler()
+    consumer = aiopg_listen.NotificationListener(aiopg_listen.connect_func(**pg_server["pg_params"]))
     consume_task = asyncio.create_task(
-        consumer.consume({"inactive_1": processor_1.process, "inactive_2": processor_2.process}, notification_timeout=1)
+        consumer.run({"inactive_1": handler_1.handle, "inactive_2": handler_2.handle}, notification_timeout=1)
     )
 
     await asyncio.sleep(1.5)
     await cancel_and_wait(consume_task)
 
-    assert processor_1.notifications == [aiopg_listen.Timeout("inactive_1")]
-    assert processor_2.notifications == [aiopg_listen.Timeout("inactive_2")]
+    assert handler_1.notifications == [aiopg_listen.Timeout("inactive_1")]
+    assert handler_2.notifications == [aiopg_listen.Timeout("inactive_2")]
 
 
 async def test_one_active_channel_and_one_passive_channel(pg_server: Dict[str, Any]) -> None:
-    active_processor = Processor()
-    inactive_processor = Processor()
-    consumer = aiopg_listen.NotificationConsumer(aiopg_listen.connect_func(**pg_server["pg_params"]))
+    active_handler = TestHandler()
+    inactive_handler = TestHandler()
+    consumer = aiopg_listen.NotificationListener(aiopg_listen.connect_func(**pg_server["pg_params"]))
     consume_task = asyncio.create_task(
-        consumer.consume(
-            {"active": active_processor.process, "inactive": inactive_processor.process}, notification_timeout=1
-        )
+        consumer.run({"active": active_handler.handle, "inactive": inactive_handler.handle}, notification_timeout=1)
     )
 
     async with aiopg.connect(**pg_server["pg_params"]) as connection, connection.cursor() as cursor:
@@ -56,21 +54,21 @@ async def test_one_active_channel_and_one_passive_channel(pg_server: Dict[str, A
 
     await cancel_and_wait(consume_task)
 
-    assert active_processor.notifications == [
+    assert active_handler.notifications == [
         aiopg_listen.Notification("active", "1"),
         aiopg_listen.Notification("active", "2"),
     ]
-    assert inactive_processor.notifications == [
+    assert inactive_handler.notifications == [
         aiopg_listen.Timeout("inactive"),
     ]
 
 
 async def test_two_active_channels(pg_server: Dict[str, Any]) -> None:
-    processor_1 = Processor()
-    processor_2 = Processor()
-    consumer = aiopg_listen.NotificationConsumer(aiopg_listen.connect_func(**pg_server["pg_params"]))
+    handler_1 = TestHandler()
+    handler_2 = TestHandler()
+    consumer = aiopg_listen.NotificationListener(aiopg_listen.connect_func(**pg_server["pg_params"]))
     consume_task = asyncio.create_task(
-        consumer.consume({"active_1": processor_1.process, "active_2": processor_2.process}, notification_timeout=1)
+        consumer.run({"active_1": handler_1.handle, "active_2": handler_2.handle}, notification_timeout=1)
     )
     await asyncio.sleep(0.1)
 
@@ -83,21 +81,21 @@ async def test_two_active_channels(pg_server: Dict[str, Any]) -> None:
 
     await cancel_and_wait(consume_task)
 
-    assert processor_1.notifications == [
+    assert handler_1.notifications == [
         aiopg_listen.Notification("active_1", "1"),
         aiopg_listen.Notification("active_1", "4"),
     ]
-    assert processor_2.notifications == [
+    assert handler_2.notifications == [
         aiopg_listen.Notification("active_2", "2"),
         aiopg_listen.Notification("active_2", "3"),
     ]
 
 
-async def test_consume_policy_last(pg_server: Dict[str, Any]) -> None:
-    processor = Processor(delay=0.1)
-    consumer = aiopg_listen.NotificationConsumer(aiopg_listen.connect_func(**pg_server["pg_params"]))
+async def test_listen_policy_last(pg_server: Dict[str, Any]) -> None:
+    handler = TestHandler(delay=0.1)
+    consumer = aiopg_listen.NotificationListener(aiopg_listen.connect_func(**pg_server["pg_params"]))
     consume_task = asyncio.create_task(
-        consumer.consume({"simple": processor.process}, policy=aiopg_listen.ConsumePolicy.LAST, notification_timeout=1)
+        consumer.run({"simple": handler.handle}, policy=aiopg_listen.ListenPolicy.LAST, notification_timeout=1)
     )
     await asyncio.sleep(0.1)
 
@@ -108,16 +106,16 @@ async def test_consume_policy_last(pg_server: Dict[str, Any]) -> None:
 
     await cancel_and_wait(consume_task)
 
-    assert processor.notifications == [
+    assert handler.notifications == [
         aiopg_listen.Notification("simple", "0"),
         aiopg_listen.Notification("simple", "9"),
     ]
 
 
-async def test_consume_policy_all(pg_server: Dict[str, Any]) -> None:
-    processor = Processor(delay=0.05)
-    consumer = aiopg_listen.NotificationConsumer(aiopg_listen.connect_func(**pg_server["pg_params"]))
-    consume_task = asyncio.create_task(consumer.consume({"simple": processor.process}, notification_timeout=1))
+async def test_listen_policy_all(pg_server: Dict[str, Any]) -> None:
+    handler = TestHandler(delay=0.05)
+    consumer = aiopg_listen.NotificationListener(aiopg_listen.connect_func(**pg_server["pg_params"]))
+    consume_task = asyncio.create_task(consumer.run({"simple": handler.handle}, notification_timeout=1))
     await asyncio.sleep(0.1)
 
     async with aiopg.connect(**pg_server["pg_params"]) as connection, connection.cursor() as cursor:
@@ -127,17 +125,17 @@ async def test_consume_policy_all(pg_server: Dict[str, Any]) -> None:
 
     await cancel_and_wait(consume_task)
 
-    assert processor.notifications == [aiopg_listen.Notification("simple", str(i)) for i in range(10)]
+    assert handler.notifications == [aiopg_listen.Notification("simple", str(i)) for i in range(10)]
 
 
 async def test_failed_to_connect(pg_server: Dict[str, Any]) -> None:
     async def connect() -> aiopg.Connection:
         raise RuntimeError("Failed to connect")
 
-    processor = Processor()
-    consumer = aiopg_listen.NotificationConsumer(connect)
-    consume_task = asyncio.create_task(consumer.consume({"simple": processor.process}, notification_timeout=1))
+    handler = TestHandler()
+    consumer = aiopg_listen.NotificationListener(connect)
+    consume_task = asyncio.create_task(consumer.run({"simple": handler.handle}, notification_timeout=1))
     await asyncio.sleep(1.5)
     await cancel_and_wait(consume_task)
 
-    assert processor.notifications == [aiopg_listen.Timeout("simple")]
+    assert handler.notifications == [aiopg_listen.Timeout("simple")]