+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## v0.0.2 (2021-07-25)
 
 * Add `aiopg_listen.connect_func` to simplify initialization [#5](https://github.com/Pliner/aiopg-listen/pull/5)
-
+* Rename consumer to listener [#6](https://github.com/Pliner/aiopg-listen/pull/6)
 
 ## v0.0.1 (2021-07-25)